start to make basic bean validation for field

+++ src/main/java/com/example/PrisonManagement/Repository/InfirmaryRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface InfirmaryRepository
-        extends JpaRepository<Infirmary, Long> {
+        extends JpaRepository<Infirmary, Integer> {
 
 }
